Introducing modules main= executable gnu_sort= sorting functionality
